Refactor reindex_rethrottle
